requirements.txt
django<=4.1
djangorestframework
django-filter
Pillow
python-dotenv
django-extensions